feat: audit logging and message logging enhancement
fix: sap-jmxdumper exception during initialization

fix: ChannelMonitoringLogger throws exception during initialization
