+++ build.gradle
@@ -25,7 +25,7 @@ project.ext {
   javaAgentVersion = '7.4.0'
   
   // Global implementation version - update this for all releases
-  implementationVersion = '1.5.0'
+  implementationVersion = '1.6.0'
 
   // Aligned with minimum Java major version supported by latest Java Agent
   javaVersion = JavaVersion.VERSION_1_8
