+++ sap-log-audit/build.gradle
@@ -8,18 +8,23 @@ dependencies {
    // Example:
    // implementation 'javax.servlet:servlet-api:2.5'
 
-   // New Relic Labs Java Agent dependencies
-   implementation fileTree(include: ['*.jar'], dir: '../libs')
+   // New Relic Java Agent dependencies
+   implementation 'com.newrelic.agent.java:newrelic-agent:7.4.0'
+   implementation 'com.newrelic.agent.java:newrelic-api:7.4.0'
    implementation fileTree(include: ['*.jar'], dir: 'lib')
+   implementation fileTree(include: ['*.jar'], dir: '../libs')
+   implementation fileTree(include: ['*.jar'], dir: '../test-lib')
 }
 
 jar {
+	
   manifest {
     attributes 'Implementation-Title': 'com.newrelic.instrumentation.labs.sap-log-audit'
     attributes 'Implementation-Vendor': 'New Relic Labs'
     attributes 'Implementation-Vendor-Id': 'com.newrelic.labs'
     attributes 'Implementation-Version': project.parent.ext.implementationVersion
-  }
+	attributes 'Agent-Class': 'com.newrelic.instrumentation.labs.sap.auditlogging.AuditLoggingPreMain'
+ }
 }
 
 verifyInstrumentation {
