+++ sap-log-audit/src/main/java/com/newrelic/instrumentation/labs/sap/auditlogging/NRLabsFormatter.java
@@ -1,0 +1,18 @@
+package com.newrelic.instrumentation.labs.sap.auditlogging;
+
+import java.util.logging.Formatter;
+import java.util.logging.LogRecord;
+
+/**
+ * Simple formatter for SAP audit logging
+ * Adapted from the sap-channel-monitoring implementation
+ * 
+ * @author gsidhwani
+ */
+public class NRLabsFormatter extends Formatter {
+
+    @Override
+    public String format(LogRecord record) {
+        return record.getMessage() + "\n";
+    }
+}
